Clear cache and remove unwanted files
