+++ back/app/controllers/api/v1/auth/sessions_controller.rb
@@ -1,2 +1,3 @@
-class Api::V1::Auth::SessionsController < DeviseTokenAuth::ApplicationController
+class Api::V1::Auth::SessionsController < DeviseTokenAuth::SessionsController
+  include ActionController::Cookies
 end
