+++ back/config/routes.rb
@@ -1,6 +1,7 @@
 Rails.application.routes.draw do
   mount_devise_token_auth_for 'User', at: 'auth', controllers: {
-    registrations: 'api/v1/auth/registrations'
+    registrations: 'api/v1/auth/registrations',
+    sessions: 'api/v1/auth/sessions'
   }
   namespace :api do
     namespace :v1 do
