+++ front/src/components/User/SignUpForm.jsx
@@ -24,19 +24,19 @@ export function SignUpForm() {
   const sendDataToAPI = async (data) => {
     const { confirmPassword, ...filteredData } = data;
     try {
-      await signUpRequest(filteredData);
+      const response = await signUpRequest(filteredData);
       router.push("/post");
     } catch (error) {
-    // console.log(error);
-    if (error.email) {
-        setError("email", { type: "manual", message: error.email });
-      } else if (error.password) {
-        setError("password", { type: "manual", message: error.password });
-      } else if (error.username) {
-        setError("username", { type: "manual", message: error.username });
+      console.log(error);
+    if (error.errors?.email) {
+        setError("email", { type: "manual", message: error.errors.email[0] });
+      } else if (error.errors?.password) {
+        setError("password", { type: "manual", message: error.errors.password[0] });
+      } else if (error.errors?.username) {
+        setError("username", { type: "manual", message: error.errors.username[0] });
       } else {
         // 他の特定フィールドでのエラーがない場合、フォーム全体に対するエラーメッセージを設定
-        setFormError(error.general);
+        setFormError("エラーが発生しました。再度お試しください。");
       }
     }
   };
@@ -76,19 +76,7 @@ export function SignUpForm() {
       <Button variant="outlined" startIcon={<GoogleIcon />} fullWidth >
       SIGN UP WITH GOOGLE
       </Button>
-      <Divider variant="fullWidth" sx={{
-        width: "100%",
-        "&::before, &::after": {
-          borderTop: '1px solid ${theme.palette.divider}',
-          content: '""',
-          flexGrow: 1
-        },
-        "& .MuiDivider-wrapper": {
-          padding: "0 10px",
-          color: '${theme.palette.text.disabled}'
-        }
-      }}
-      >or</Divider>
+      <Divider variant="fullWidth">or</Divider>
 
       <Box
       sx={{
